+++ Sources/KeyboardKit/Actions/StandardSecondaryCalloutActionProvider.swift
@@ -1,0 +1,23 @@
+//
+//  StandardSecondaryCalloutActionProvider.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2021-01-06.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This provider is used by default, and provides the standard
+ secondary callout actions for the current locale, if any.
+ 
+ You can inherit and customize this class to create your own
+ provider that builds on this foundation.
+ */
+open class StandardSecondaryCalloutActionProvider: SecondaryCalloutActionProvider {
+    
+    open func secondaryCalloutActions(for action: KeyboardAction, in context: KeyboardContext) -> [KeyboardAction] {
+        return []
+    }
+}