+++ Sources/KeyboardKit/Input/StandardKeyboardInputSetProvider.swift
@@ -1,5 +1,5 @@
 //
-//  StandardKeyboardInputProvider.swift
+//  StandardKeyboardInputSetProvider.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-12-01.
@@ -19,7 +19,7 @@ import Foundation
  They are not part of the protocol, but you can use them and
  extend them in your own subclasses.
  */
-open class StandardKeyboardInputProvider: KeyboardInputProvider {
+open class StandardKeyboardInputSetProvider: KeyboardInputSetProvider {
     
     public init() {}